Add signal multiplier 0.5 and 0.25 for FW 2.5/3.0
* Change signal multiplier type to double
* Make sure set_config rounds to int for values 1, 2, 3 for
compatibility with lower FWs
* Add note about intrinsics/metadata being available earlier in FW
2.5/3.0
* Add references to 4096x5
* Specify which to_string for mode and ts_mode

+++ ouster_client/include/ouster/types.h
@@ -238,7 +238,7 @@ struct sensor_config {
      * Multiplier for signal strength of sensor. See the sensor docs for more
      * details on usage.
      */
-    optional<int> signal_multiplier;
+    optional<double> signal_multiplier;
 
     /**
      * The nmea polarity for the sensor to use.
