+++ python/src/ouster/client/_client.pyi
@@ -498,7 +498,7 @@ class SensorConfig:
     operating_mode: Optional[OperatingMode]
     multipurpose_io_mode: Optional[MultipurposeIOMode]
     azimuth_window: Optional[tuple]
-    signal_multiplier: Optional[int]
+    signal_multiplier: Optional[float]
     sync_pulse_out_angle: Optional[int]
     sync_pulse_out_pulse_width: Optional[int]
     nmea_in_polarity: Optional[Polarity]
